Fix mismatched IMU mobile base link / topic
As of PR #95, the Stretch base IMU xacro is included every time
a URDF is generated. This includes the URDF link corresponding
to the mobile base IMU. This link has been called 'base_imu' in
the past, but the topic from the driver refers to a nonexistant
link called 'imu_mobile_base'. To fix the mismatch, the xacro's
link is renamed from 'base_imu' to 'imu_mobile_base'.