Reduce console clutter in driver launch
A few nodes are launched alongside Stretch Driver with the driver
launch file is called. This includes robot_state_publisher, which
has started to print debug info to the console. To reduce the
clutter in the console, these debug messages are silenced. Also,
a parameter warning in Stretch Driver is fixed and some info level
logs are changed to debug level.